+++ src/common.rs
@@ -451,3 +451,8 @@ impl LabelRecog {
         }
     }
 }
+
+pub fn present_error(msg: String) {
+    eprintln!("[Error] {msg}");
+    std::process::exit(1)
+}